+++ src/App.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import './App.css';
 import { HashRouter as Router, Route, Switch } from 'react-router-dom';
-import { Homepage } from './components/main/Homepage';
-import { ChooseRoom } from './components/main/ChooseRoom';
-import { ChooseDate } from './components/main/ChooseDate';
-import { ChooseTime } from './components/main/ChooseTime';
-import { BookingSummary } from './components/main/BookingSummary';
+import {
+  Homepage,
+  ChooseRoom,
+  ChooseDate,
+  ChooseTime,
+  BookingSummary
+} from './components/main';
 
 const App = () => {
   return (
