+++ src/components/BookingSummary/index.js
@@ -1,0 +1,3 @@
+import { Modals } from './Modals';
+
+export { Modals };
