+++ src/components/ChooseDate/index.js
@@ -1,0 +1,3 @@
+import { DatePicker } from './DatePicker';
+
+export { DatePicker };
