+++ src/components/ChooseRoom/index.js
@@ -1,0 +1,3 @@
+import { RoomList } from './RoomList';
+
+export { RoomList };
