+++ src/components/complement/index.js
@@ -1,0 +1,4 @@
+import { Loading } from './Loading';
+import { NavBar } from './NavBar';
+
+export { Loading, NavBar };
