+++ src/components/main/index.js
@@ -1,0 +1,7 @@
+import { Homepage } from './Homepage';
+import { ChooseRoom } from './ChooseRoom';
+import { ChooseDate } from './ChooseDate';
+import { ChooseTime } from './ChooseTime';
+import { BookingSummary } from './BookingSummary';
+
+export { Homepage, ChooseRoom, ChooseDate, ChooseTime, BookingSummary };
