+++ src/functions/index.js
@@ -1,0 +1,6 @@
+import { redirectToLogin } from './redirectToLogin';
+import { shortlistBookings } from './shortlistBookings';
+import { validateTime } from './validateTime';
+import { getCurrentHour } from './getCurrentHour';
+
+export { redirectToLogin, shortlistBookings, validateTime, getCurrentHour };
